+++ packages/utils/src/index.ts
@@ -1,5 +1,6 @@
-export {activeElement} from './activeElement';
-export {contains} from './contains';
-export {getWindow} from './getWindow';
+export * from './activeElement';
+export * from './contains';
+export * from './getWindow';
 export * from './is';
-export {getNodeName, isNode} from './node';
+export * from './node';
+export * from './userAgent';
